Time: 2 ms (54.17%), Space: 6.5 MB (31.25%) - LeetHub

+++ 0062-unique-paths/0062-unique-paths.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
-long long uniquePaths(int m, int n) {
+int  uniquePaths(int m, int n) {
     if (m <= 0 || n <= 0) return 0;
 
-    long long dp[m][n];
+    int dp[m][n];
 
     // Initialize the first row and first column to 1
     for (int i = 0; i < m; ++i) {
